Add calculateFee to ThorchainEngine
